Revert "Copy the calendar url rather than the archive url"
This reverts commit c763a0527e5ed497733a5f567dcd018a82177201.

Paul Koerbin writes:

After talking with Dianne and considering the probable (possible?) release of the new web archive delivery system I think it would be best if the ‘Copy Archive Link’ was actually the same as the displayed link – that is, it should be the link with date stamp pointing to the actual instance rather than with the wildcard asterisk pointing to the calendar. If the new delivery system is implemented it makes more sense that this link should be to the dated instance.